Initial checkin of Linux files.

+++ src/base/lib/linux-headers/asm-i386/types.h
@@ -16,8 +16,8 @@ typedef unsigned char __u8;
 typedef __signed__ short __s16;
 typedef unsigned short __u16;
 
-typedef __signed__ int __s32;
-typedef unsigned int __u32;
+typedef __signed__ /* int modified to match stdint.h */ long __s32;
+typedef unsigned /* int modified to match stdint.h */ long __u32;
 
 #if defined(__GNUC__) && !defined(__STRICT_ANSI__)
 typedef __signed__ long long __s64;
